pyproject.toml
[tool.poetry]
name = "langboard"
version = "0.1.0"
description = ""
authors = ["Yamon <battlecruser@yamon.io>"]
readme = "README.md"
packages = [{include = "langboard", from = "src/backend"}]
include = ["src/backend/langboard/*", "src/backend/langboard/**/*"]


[tool.poetry.scripts]
langboard = 'langboard.CLI:execute'


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.112.0"
pydantic = "^2.8.2"
socketify = "^0.0.28"
routes = "^2.5.1"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
sqlmodel = "^0.0.22"
bcrypt = "^4.2.0"
cryptocode = "^0.1"
pyjwt = "^2.9.0"
redis = "^5.1.0"
boto3 = "^1.35.37"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.5.5"
pytest = "^8.3.2"
rich = "^13.8.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
boto3-stubs = {version = "1.35.37", extras = ["s3"]}


[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = ["Test*.py"]
console_output_style = "progress"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = ["ignore::DeprecationWarning", "ignore::ResourceWarning"]
log_cli = true
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


[tool.coverage.report]
omit = [
    "src/backend/langboard/core/bootstrap/CLIHelpFormatter.py",
    "src/backend/langboard/core/bootstrap/CLIOptions.py",
    "src/backend/langboard/core/bootstrap/CLIRichParser.py",
    "src/backend/langboard/core/bootstrap/WebSocketOptions.py",
    "src/backend/langboard/core/db/ModelCreator.py",
    "src/backend/langboard/core/filter/BaseFilter.py",
    "src/backend/langboard/core/routing/WebSocket.py",
    "src/backend/langboard/core/routing/BaseMiddleware.py",
    "src/backend/langboard/App.py",
    "src/backend/langboard/CLI.py",
    "src/backend/langboard/Constants.py",
    "src/backend/langboard/Loader.py",
    "**/__init__.py",
]


[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
disable_error_code = ["type-var"]


[tool.ruff]
line-length = 120
exclude = ["migrations"]
lint.extend-select = ["I"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"


[tool.ruff.lint.isort]
lines-after-imports = 2
no-lines-before = ["future", "standard-library", "third-party", "first-party", "local-folder"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
